+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Fix up deployment scripts
 * add User schema
 * add Credential schema
+* add undeploy and upsert-deploy to kubernetes tasks